draft of default texture collections support
Specified by the optional defaultCollections property in the textures object of the game config. Value is a list of texture collection strings, where each string is in the same path format as the components stored in the worldspawn attribute, e.g.

    "defaultCollections": [ "textures/common", "textures/ctf" ]

It's arguable whether the "textures/" prefix should be required given that the textures object also specifies the package root ... ?

Any textures specified here will be automatically added to the initial loaded collections, and cannot be removed from the list of loaded collections.

Probably we would need to highlight/badge them in some fashion in the texture collections UI to indicate that they can't be removed.

Also need some new tests.